Working on conversion to logical pixels.

+++ examples/LiveShaderEditing/live_shader_editing.cpp
@@ -81,7 +81,7 @@ public:
   }
 
   void resized() override {
-    int center = width() / 2;
+    float center = width() / 2.0f;
     int shapes_width = std::min(center, height());
     shapes_.setBounds((center - shapes_width) / 2, (height() - shapes_width) / 2, shapes_width, shapes_width);
     shader_editor_.setBounds(center, 0, width() - center, height());
